+++ index.php
@@ -15,8 +15,8 @@ $searchSqlParts = [];
 if ($_SERVER["REQUEST_METHOD"] === "POST") {
     if (isset($_POST["deleteComment"])) {
         $commentID = isset($_POST["commentID"]) ? $_POST["commentID"] : '';
-        $deleteStmt = $conn->prepare("DELETE FROM comments WHERE id = ?");
-        $deleteStmt->bind_param("i", $commentID);
+        $deleteStmt = $conn->prepare("DELETE FROM comments WHERE id = ? AND user_id = ?");
+        $deleteStmt->bind_param("ii", $commentID , $_SESSION["id"]);
         $deleteStmt->execute();
     }
 }
